@@ -1,6 +1,6 @@
 {
   "name": "angular-webpack-visualstudio",
-  "version": "4.0.3",
+  "version": "5.0.0",
   "description": "An Angular VS template",
   "main": "wwwroot/index.html",
   "author": "",
@@ -23,68 +23,72 @@
     "lint": "tslint ./angularApp"
   },
   "dependencies": {
-    "bootstrap": "4.1.0",
-    "@ngrx/effects": "^5.2.0",
-    "@ngrx/store": "^5.2.0",
-    "@ngrx/store-devtools": "5.2.0",
-    "@angular/animations": "5.2.10",
-    "@angular/common": "5.2.10",
-    "@angular/compiler": "5.2.10",
-    "@angular/compiler-cli": "5.2.10",
-    "@angular/core": "5.2.10",
-    "@angular/forms": "5.2.10",
-    "@angular/http": "5.2.10",
-    "@angular/platform-browser": "5.2.10",
-    "@angular/platform-browser-dynamic": "5.2.10",
-    "@angular/platform-server": "5.2.10",
-    "@angular/router": "5.2.10",
-    "@angular/upgrade": "5.2.10",
+    "@angular-devkit/core": "0.6.0",
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/platform-server": "6.0.0",
+    "@angular/router": "6.0.0",
+    "@angular/upgrade": "6.0.0",
     "@aspnet/signalr": "^1.0.0-preview2-final",
+    "@ngrx/effects": "6.0.0-beta.1",
+    "@ngrx/store": "6.0.0-beta.1",
+    "@ngrx/store-devtools": "6.0.0-beta.1",
     "angular-auth-oidc-client": "4.1.0",
+    "bootstrap": "4.1.1",
     "core-js": "2.5.5",
     "ie-shim": "0.1.0",
     "msgpack5": "4.0.2",
-    "rxjs": "5.5.10",
+    "npm-check-updates": "^2.14.2",
+    "rxjs": "6.1.0",
+    "rxjs-compat": "6.1.0",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@ngtools/webpack": "^1.10.2",
-    "@types/jasmine": "^2.8.6",
-    "@types/node": "^9.6.6",
-    "angular-router-loader": "^0.8.2",
+    "@ngtools/webpack": "^6.0.0",
+    "@types/jasmine": "^2.8.7",
+    "angular-router-loader": "0.8.5",
+    "@types/node": "^10.0.3",
     "angular2-template-loader": "^0.6.2",
-    "awesome-typescript-loader": "^3.4.1",
+    "awesome-typescript-loader": "^5.0.0",
     "clean-webpack-plugin": "^0.1.19",
-    "codelyzer": "^4.2.1",
+    "codelyzer": "^4.3.0",
     "concurrently": "^3.5.1",
     "copy-webpack-plugin": "^4.5.1",
-    "css-loader": "^0.28.10",
+    "css-loader": "^0.28.11",
     "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^3.0.6",
-    "jasmine-core": "^2.99.1",
+    "html-webpack-plugin": "^3.2.0",
+    "jasmine-core": "3.1.0",
     "jquery": "^3.3.1",
     "json-loader": "^0.5.7",
-    "karma": "^2.0.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-jasmine": "^1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.0.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "^0.0.32",
-    "karma-webpack": "^2.0.9",
-    "node-sass": "^4.8.3",
+    "karma-webpack": "^3.0.0",
+    "node-sass": "^4.9.0",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.6.2",
-    "sass-loader": "^6.0.7",
+    "sass-loader": "^7.0.1",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.20.3",
-    "tslint": "^5.9.1",
+    "style-loader": "^0.21.0",
+    "tslint": "^5.10.0",
     "tslint-loader": "^3.6.0",
-    "typescript": "2.6.2",
-    "uglifyjs-webpack-plugin": "^1.2.3",
+    "typescript": "2.8.3",
+    "uglifyjs-webpack-plugin": "^1.2.5",
     "url-loader": "^1.0.1",
-    "webpack": "3.11.0",
+    "webpack": "^4.6.0",
     "webpack-bundle-analyzer": "^2.11.1",
-    "webpack-dev-server": "^3.1.1"
+    "webpack-dev-server": "^3.1.4",
+    "webpack-cli": "2.1.2"
   },
   "-vs-binding": {
     "ProjectOpened": [
