@@ -1,6 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
-import { Observable } from 'rxjs/Observable';
+import { Subscription ,  Observable } from 'rxjs';
 import { Store } from '@ngrx/store';
 import { NewsState } from '../store/news.state';
 import * as NewsActions from '../store/news.action';
@@ -28,12 +27,15 @@ export class NewsComponent implements OnInit, OnDestroy {
         private store: Store<any>,
         private oidcSecurityService: OidcSecurityService
     ) {
-        this.newsState$ = this.store.select<NewsState>(state => state.news.newsitems);
+        this.newsState$ = this.store.select<NewsState>(state => state.news);
 
-        this.store.select<NewsState>(state => state.news.newsitems).subscribe((o: NewsState) => {
-            this.newsItems = o.newsItems;
+        this.store.select<NewsState>(state => state.news).subscribe((o: NewsState) => {
+            console.log('event');
+            console.log(o);
+            this.newsItems = o.news.newsItems;
         });
 
+        console.log(this.newsItems);
         this.newsItem = new NewsItem();
         this.newsItem.AddData('', '', this.author, this.group);
     }
@@ -45,6 +47,7 @@ export class NewsComponent implements OnInit, OnDestroy {
     }
 
     public join(): void {
+        console.log('join');
         this.store.dispatch(new NewsActions.JoinGroupAction(this.group));
     }
 
@@ -57,6 +60,7 @@ export class NewsComponent implements OnInit, OnDestroy {
             (isAuthorized: boolean) => {
                 this.isAuthorized = isAuthorized;
                 if (this.isAuthorized) {
+                    console.log('this.store.dispatch(new NewsActions.SelectAllGroupsAction()');
                     this.store.dispatch(new NewsActions.SelectAllGroupsAction());
                 }
             });
