@@ -2,8 +2,10 @@ import { NewsState } from './news.state';
 import * as newsAction from './news.action';
 
 export const initialState: NewsState = {
-    newsItems: [],
-    groups: ['IT', 'global', 'sport']
+    news: {
+        newsItems: [],
+        groups: ['IT', 'global', 'sport']
+    }
 };
 
 export function newsReducer(state = initialState, action: newsAction.Actions): NewsState {
@@ -11,34 +13,49 @@ export function newsReducer(state = initialState, action: newsAction.Actions): N
 
         case newsAction.RECEIVED_GROUP_JOINED:
             return Object.assign({}, state, {
-                groups: (state.groups.indexOf(action.group) > -1) ? state.groups : state.groups.concat(action.group)
+                news: {
+                    newsItems: state.news.newsItems,
+                    groups: (state.news.groups.indexOf(action.group) > -1) ? state.news.groups : state.news.groups.concat(action.group)
+                }
             });
 
         case newsAction.RECEIVED_NEWS_ITEM:
             return Object.assign({}, state, {
-                newsItems: state.newsItems.concat(action.newsItem),
+                news: {
+                    newsItems: state.news.newsItems.concat(action.newsItem),
+                    groups: state.news.groups
+                }
             });
 
         case newsAction.RECEIVED_GROUP_HISTORY:
             return Object.assign({}, state, {
-                newsItems: action.newsItems
+                news: {
+                    newsItems: action.newsItems,
+                    groups: state.news.groups
+                }
             });
 
         case newsAction.RECEIVED_GROUP_LEFT:
             const data = [];
-            for (const entry of state.groups) {
+            for (const entry of state.news.groups) {
                 if (entry !== action.group) {
                     data.push(entry);
                 }
             }
             console.log(data);
             return Object.assign({}, state, {
-                groups: data
+                news: {
+                    newsItems: state.news.newsItems,
+                    groups: data
+                }
             });
 
         case newsAction.SELECTALL_GROUPS_COMPLETE:
             return Object.assign({}, state, {
-                groups: action.groups
+                news: {
+                    newsItems: state.news.newsItems,
+                    groups: action.groups
+                }
             });
 
         default:
