@@ -1,7 +1,6 @@
-﻿import 'jquery/src/jquery';
+﻿import 'jquery/dist/jquery';
 import 'bootstrap/dist/css/bootstrap.css';
 import 'bootstrap/dist/js/bootstrap.bundle.min.js';
-
 
 import '../favicon.ico';
 
