@@ -1,7 +1,7 @@
 ﻿// Karma configuration for Angular testing
 
 module.exports = function (config) {
-    var webpackConfig = require('./config//webpack.test.js');
+    var webpackConfig = require('./config/webpack.test.js');
 
     var configuration = {
 
@@ -21,6 +21,10 @@ module.exports = function (config) {
         exclude: [
         ],
 
+        client: {
+            clearContext: false
+        },
+
         // preprocess matching files before serving them to the browser
         // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
         preprocessors: {
@@ -38,7 +42,7 @@ module.exports = function (config) {
         // test results reporter to use
         // possible values: 'dots', 'progress'
         // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-        reporters: ['spec'],
+        reporters: ['spec', 'kjhtml'],
 
 
         // web server port
@@ -64,7 +68,7 @@ module.exports = function (config) {
 
         // Continuous Integration mode
         // if true, Karma captures browsers, runs the tests and exits
-        singleRun: true
+        singleRun: false
 
     };
 
